+++ test/config_test2.js
@@ -1,13 +1,13 @@
 profiles = {
   default: {
-  	Environment: {},
-  	PIRMotionDetector: {
-          pin: 17
-        },
-	SlackNotifier: {
-	  name: "My Slack channel",
-	  key: "https://hooks.slack.com/services/<MySlackURL>"
-	}
+    Environment: {},
+    PIRMotionDetector: {
+      pin: 17
+    },
+    SlackNotifier: {
+      name: "My Slack channel",
+      key: "https://hooks.slack.com/services/<MySlackURL>"
+    }
   }
 }