+++ test/config_test3.js
@@ -1,0 +1,12 @@
+profiles = {
+  default: {
+    Environment: {},
+    MotionDetector: {
+    },
+    BaseNotifier: {
+    }
+  }
+}
+
+exports.profiles = profiles;
+exports.default = profiles.default;