+++ test/config_test4.js
@@ -1,0 +1,13 @@
+profiles = {
+  default: {
+    Environment: {},
+    MotionDetector: {
+      name: "MD 1"
+    },
+    BaseNotifier: {
+    }
+  }
+}
+
+exports.profiles = profiles;
+exports.default = profiles.default;